requirements.txt
flask
requests
python-dotenv
aiohttp
python-dotenv
